chore(board): export code related with jackson mapper config in driving adapter test code
